add executable name to filename of output remove static_epoch_xname variable (replaced by exec name) backtrace print on unw failure csprof main makefile uses private papi MPI_Init override to get rank for csprof (add to output filename when appropriate) add support for last sample address, and incrementing of counters cleanup sigaction signature (add const when appropriate) catamount/cnl gethostid returns mpirank papi sampler option to not use callstack papi_debug set cct cleanup some compile warnings stack troll subtraction upgrade to long

+++ branches/csprof-catamount/BetterUnwind/stack_troll.c
@@ -29,7 +29,7 @@ unsigned int stack_troll(char **sp, unsigned int *ra_pos){
     char *s, *e;
     if (! find_enclosing_function_bounds_v(*tmp_sp,&s,&e,SILENT)){
       PMSG(TROLL,"found valid address %p at sp = %p",*tmp_sp,tmp_sp);
-      *ra_pos = (unsigned int)tmp_sp - (unsigned int)sp;
+      *ra_pos = (unsigned long) tmp_sp - (unsigned long) sp;
       return 1;
     }
     tmp_sp++;
